refactor(media): renamed and moved to admin folder
